pf2e attack roll and setup
