+++ module.json
@@ -28,6 +28,15 @@
 					"minimum": "2.0.2",
 					"verified": "2.0.2"
 				}
+			},
+			{
+				"id": "pf2e",
+				"type": "system",
+				"manifest": "https://github.com/foundryvtt/pf2e/releases/download/4.4.0/system.json",
+				"compatibility": {
+					"minimum": "4.4.0",
+					"verified": "4.4.0"
+				}
 			}
 		]
 	},
